added introduction_table
