+++ src/main/java/com/example/hotelbookingweb/entities/RoomEntity.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 @AllArgsConstructor
-
 public class RoomEntity {
     private int id;
+    private int introduction_id;
     private int roomNum;
     private int floor;
+    private int price;
     private String roomType;
     private int roomCapacity;
     private boolean isAvailable;
-
 }
