+++ src/main/java/com/example/hotelbookingweb/input_form/BookingForm.java
@@ -1,11 +1,13 @@
 package com.example.hotelbookingweb.input_form;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import java.time.LocalDate;
 
 
 @Data
+@AllArgsConstructor
 public class BookingForm {
 
     private String name;
@@ -15,12 +17,4 @@ public class BookingForm {
     private LocalDate checkInDate;
     private LocalDate checkOutDate;
 
-    public BookingForm(String name, int roomNum, Integer age, int gender, LocalDate checkInDate, LocalDate checkOutDate) {
-        this.name = name;
-        this.roomNum = roomNum;
-        this.age = age;
-        this.gender = gender;
-        this.checkInDate = checkInDate;
-        this.checkOutDate = checkOutDate;
-    }
 }
