+++ src/test/java/com/example/hotelbookingweb/controllers/users/RoomInfoControllerTest.java
@@ -14,8 +14,8 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.LocalDate;
 
-import static com.example.hotelbookingweb.controllers.users.RoomInfoController.CONFIRM_BOOKING_PATH;
-import static com.example.hotelbookingweb.controllers.users.RoomInfoController.ROOMS_INFORMATION_PATH;
+import static com.example.hotelbookingweb.controllers.users.BookingFormController.CONFIRM_BOOKING_PATH;
+import static com.example.hotelbookingweb.controllers.users.BookingFormController.ROOMS_INFORMATION_PATH;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
